[Refector] 쿠폰 서비스 리팩토링
* 쿠폰 facade 삭제하고 도메인 서비스에서 쿠폰 발급, 조회 로직 수행

+++ src/main/java/kr/hhplus/be/server/domain/coupon/entity/Coupon.java
@@ -13,6 +13,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -29,7 +30,7 @@ public class Coupon extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private DiscountType discountType;
 
-    private int discountAmt;
+    private BigDecimal discountAmt;
 
     private int maxCapacity;
 
@@ -43,7 +44,7 @@ public class Coupon extends BaseEntity {
     private CouponStatus status;
 
     @Builder
-    public Coupon(String name, DiscountType discountType, int discountAmt, int maxCapacity, int remainCapacity, LocalDateTime validStartedAt, LocalDateTime validEndedAt, CouponStatus status) {
+    public Coupon(String name, DiscountType discountType, BigDecimal discountAmt, int maxCapacity, int remainCapacity, LocalDateTime validStartedAt, LocalDateTime validEndedAt, CouponStatus status) {
         this.name = name;
         this.discountType = discountType;
         this.discountAmt = discountAmt;
